moved the sorting to a new module
